kovo store/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
distlib==0.3.1
Django==3.2
django-address==0.2.5
django-autoslug==1.9.8
django-cities==0.6
django-cities-light==3.8.1
django-ckeditor==6.0.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-phone-field==1.8.1
django-smart-selects==1.5.9
djangotoolbox==1.8.0
filelock==3.0.12
Flask==2.0.1
idna==2.10
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
progressbar2==3.53.1
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.58.0
swapper==1.1.2.post1
tqdm==4.61.0
Unidecode==1.2.0
urllib3==1.26.6
virtualenv==20.4.4
Werkzeug==2.0.1
gunicorn==20.1.0
whitenoise==5.3.0

